:bricks: Get a user pic
