+++ src/routes/picturesRoute.ts
@@ -2,7 +2,7 @@ import { Router } from "express";
 import multer from "multer";
 import { validateJWT } from "../middlewares/tokenMiddleware";
 import multerConfig from "../multerConfig/multer";
-import * as uploadController from "../controllers/uploadControllers";
+import * as picturesControllers from "../controllers/picturesControllers";
 
 const uploadRoute = Router();
 
@@ -11,7 +11,9 @@ uploadRoute.use(validateJWT);
 uploadRoute.post(
   "/profilePic",
   multer(multerConfig).single("profilePic"),
-  uploadController.insertProfilePic
+  picturesControllers.insertProfilePic
 );
+
+uploadRoute.get("/profilePic", picturesControllers.getProfilePic);
 
 export default uploadRoute;
